+++ installer/installer.py
@@ -34,6 +34,9 @@ available.
 from argparse import ArgumentParser
 import collections
 import grp
+import html
+import inspect
+import io
 import os
 from pathlib import Path
 from platform import uname
@@ -48,6 +51,7 @@ from typing import (
     Callable,
     Dict,
     Iterable,
+    List,
     NoReturn,
     TextIO,
     Tuple,
@@ -87,7 +91,8 @@ from semantic_version import Version
 # =============================================================================
 
 MINIMUM_DOCKER_COMPOSE_VERSION = Version("2.0.0")
-EXIT_FAILURE = 1
+EXIT_FAILURE = 1  # Unexpected error
+EXIT_USER = 2  # User error e.g. bad command, misconfiguration, CTRL-C
 
 
 class HostPath:
@@ -256,9 +261,9 @@ class Ports:
     CRATEWEB = "8000"
     RABBITMQ = "5672"
     CRATE_DB_HOST = "43306"
-    RESEARCH_DB_HOST = "43307"
-    SECRET_DB_HOST = "43308"
-    SOURCE_DB_HOST = "43309"
+    DEMO_RESEARCH_DB_HOST = "43307"
+    DEMO_SECRET_DB_HOST = "43308"
+    DEMO_SOURCE_DB_HOST = "43309"
 
 
 # =============================================================================
@@ -269,6 +274,33 @@ class Ports:
 class DatabaseEngine(
     collections.namedtuple(
         "DatabaseEngine", ["description", "sqlalchemy", "django"]
+    )
+):
+    pass
+
+
+# =============================================================================
+# Databases
+# =============================================================================
+
+
+class Database(
+    collections.namedtuple(
+        "Database",
+        [
+            # Docker and external
+            "short_desc",
+            "long_desc",
+            "engine_var",
+            "host_var",
+            "port_var",
+            "name_var",
+            "user_name_var",
+            "password_var",
+            # Docker MySQL only
+            "root_password_var",
+            "host_port_var",
+        ],
     )
 ):
     pass
@@ -359,12 +391,15 @@ class Installer:
     def __init__(
         self,
         crate_root_dir: str = None,
+        light_mode: bool = False,
         update: bool = False,
         verbose: bool = False,
     ) -> None:
         self._docker = None
         self._engines = None
         self._env_dict = None
+        self._databases = None
+        self.light_mode = light_mode
         self.update = update
         self.verbose = verbose
 
@@ -377,37 +412,48 @@ class Installer:
                 "variable CRATE_INSTALLER_CRATE_ROOT_HOST_DIR"
             )
 
-            sys.exit(EXIT_FAILURE)
+            sys.exit(EXIT_USER)
+
+        self.title = "CRATE Setup"
+        self.choice_style = self.get_choice_style()
+        self.envvar_style = self.get_body_style()
+        self.error_style = self.get_text_style(Colours.RED)
+        self.info_style = self.get_body_style()
+        self.intro_style = self.get_highlight_style()
+        self.prompt_style = self.get_text_style(Colours.CYAN)
+        self.success_style = self.get_text_style(Colours.GREEN)
 
         self.setenv(InstallerEnvVar.CRATE_ROOT_HOST_DIR, crate_root_dir)
 
-        self.title = "CRATE Setup"
-        self.intro_style = Style.from_dict(
-            {
-                "span": f"{Colours.BLUE}",
-            }
-        )
-        prompt_dict = {"span": f"{Colours.CYAN}"}
+    def get_body_style(self) -> Style:
+        return self.get_text_style(self.get_body_foreground_colour())
 
-        self.prompt_style = Style.from_dict(prompt_dict)
-        self.info_style = Style.from_dict(
+    def get_body_foreground_colour(self) -> str:
+        if self.light_mode:
+            return Colours.BASE00
+
+        return Colours.BASE0
+
+    def get_highlight_style(self) -> Style:
+        if self.light_mode:
+            return self.get_text_style(
+                f"bg:{Colours.BASE2} fg:{Colours.BASE01}"
+            )
+
+        return self.get_text_style(f"bg:{Colours.BASE02} fg:{Colours.BASE1}")
+
+    @staticmethod
+    def get_text_style(style_string: str) -> Style:
+        return Style.from_dict({"span": style_string})
+
+    def get_choice_style(self) -> Style:
+        return Style.from_dict(
             {
-                "span": f"{Colours.VIOLET}",
+                "span": Colours.CYAN,
+                "name": Colours.ORANGE,
+                "desc": self.get_body_foreground_colour(),
             }
         )
-        self.error_style = Style.from_dict(
-            {
-                "span": f"{Colours.RED}",
-            }
-        )
-        self.envvar_style = Style.from_dict(
-            {
-                "span": f"{Colours.GREEN}",
-            }
-        )
-        choice_dict = prompt_dict.copy()
-        choice_dict.update(name=f"{Colours.ORANGE}")
-        self.choice_style = Style.from_dict(choice_dict)
 
     @property
     def engines(self) -> Dict[str, DatabaseEngine]:
@@ -438,6 +484,78 @@ class Installer:
                 "PostgreSQL",
                 "postgresql+psycopg2",
                 "django.db.backends.postgresql",
+            ),
+        }
+
+    @property
+    def databases(self) -> Dict[str, Database]:
+        if self._databases is None:
+            self._databases = self.get_databases()
+
+        return self._databases
+
+    def get_databases(self) -> Dict[str, Database]:
+        return {
+            "crate": Database(
+                "CRATE database",
+                "This database is used by the CRATE web application.",
+                InstallerEnvVar.CRATE_DB_ENGINE,
+                InstallerEnvVar.CRATE_DB_SERVER,
+                InstallerEnvVar.CRATE_DB_PORT,
+                DockerEnvVar.CRATE_DB_DATABASE_NAME,
+                DockerEnvVar.CRATE_DB_USER_NAME,
+                DockerEnvVar.CRATE_DB_USER_PASSWORD,
+                DockerEnvVar.CRATE_DB_ROOT_PASSWORD,
+                DockerEnvVar.CRATE_DB_HOST_PORT,
+            ),
+            "research": Database(
+                "research database",
+                (
+                    "This is where the anonymised data will be stored. More "
+                    "research databases can be added by manually editing the "
+                    "configuration files generated by CRATE after "
+                    "installation."
+                ),
+                InstallerEnvVar.RESEARCH_DATABASE_ENGINE,
+                InstallerEnvVar.RESEARCH_DATABASE_HOST,
+                InstallerEnvVar.RESEARCH_DATABASE_PORT,
+                DockerEnvVar.RESEARCH_DATABASE_NAME,
+                DockerEnvVar.RESEARCH_DATABASE_USER_NAME,
+                DockerEnvVar.RESEARCH_DATABASE_USER_PASSWORD,
+                DockerEnvVar.RESEARCH_DATABASE_ROOT_PASSWORD,
+                DockerEnvVar.RESEARCH_DATABASE_HOST_PORT,
+            ),
+            "secret": Database(
+                "secret administrative database",
+                "This database stores information like Patient ID to "
+                "Research ID mapping. "
+                "More secret databases can be added by manually editing the "
+                "configuration files generated by CRATE after installation. "
+                "There should be one secret administrative database for every "
+                "research database.",
+                InstallerEnvVar.SECRET_DATABASE_ENGINE,
+                InstallerEnvVar.SECRET_DATABASE_HOST,
+                InstallerEnvVar.SECRET_DATABASE_PORT,
+                DockerEnvVar.SECRET_DATABASE_NAME,
+                DockerEnvVar.SECRET_DATABASE_USER_NAME,
+                DockerEnvVar.SECRET_DATABASE_USER_PASSWORD,
+                DockerEnvVar.SECRET_DATABASE_ROOT_PASSWORD,
+                DockerEnvVar.SECRET_DATABASE_HOST_PORT,
+            ),
+            "source": Database(
+                "source database",
+                "This database contains the patient-identifiable "
+                "data. More source databases can be added by manually editing "
+                "the configuration files generated by CRATE after "
+                "installation.",
+                InstallerEnvVar.SOURCE_DATABASE_ENGINE,
+                InstallerEnvVar.SOURCE_DATABASE_HOST,
+                InstallerEnvVar.SOURCE_DATABASE_PORT,
+                DockerEnvVar.SOURCE_DATABASE_NAME,
+                DockerEnvVar.SOURCE_DATABASE_USER_NAME,
+                DockerEnvVar.SOURCE_DATABASE_USER_PASSWORD,
+                DockerEnvVar.SOURCE_DATABASE_ROOT_PASSWORD,
+                DockerEnvVar.SOURCE_DATABASE_HOST_PORT,
             ),
         }
 
@@ -474,6 +592,7 @@ class Installer:
         self.write_environment_variables()
 
         self.build_crate_image()
+        self.maybe_test_database_connections()
 
         # At this point the containers get created
         self.create_local_settings()
@@ -522,6 +641,64 @@ class Installer:
 
         return len(images) == 0
 
+    def maybe_test_database_connections(self) -> None:
+        # Only test external database connections, not those provided by the
+        # installer under Docker. Those won't exist yet.
+        if not self.should_create_crate_db_container():
+            self.test_database_connection(self.databases["crate"])
+
+        if not self.should_create_demo_containers():
+            self.test_database_connection(self.databases["source"])
+            self.test_database_connection(self.databases["research"])
+            self.test_database_connection(self.databases["secret"])
+
+    def test_database_connection(self, db: Database) -> None:
+        self.info(f"\nTesting connection to the {db.short_desc}...")
+        os.chdir(self.dockerfiles_host_dir())
+
+        error = io.StringIO()
+        output = io.StringIO()
+
+        try:
+            output_generator = self.run_crate_command(
+                [
+                    "crate_test_database_connection",
+                    self.get_db_url(db),
+                ],
+                stream=True,
+            )
+            for stream_type, stream_content in output_generator:
+                decoded = stream_content.decode("utf-8")
+
+                if stream_type == "stdout":
+                    output.write(decoded)
+                elif stream_type == "stderr":
+                    error.write(decoded)
+        except DockerException:
+            self.error(output.getvalue())
+
+            if self.verbose:
+                self.error(error.getvalue(), split_lines=False)
+
+                self.fail(
+                    "Failed to connect.\n"
+                    "Troubleshooting:\n"
+                    "----------------\n"
+                    "'Login timeout expired': Problem with server host name "
+                    "or firewall?\n"
+                    "'Login failed': Problem with username/password?\n"
+                    "'Data source name not found...': Incorrect DSN name or "
+                    "entry?\n\n"
+                )
+
+            self.fail(
+                "Failed to connect. "
+                "Run the installer again with --verbose to see more detail."
+            )
+
+        self.info(output.getvalue().strip())
+        self.success("OK")
+
     def start(self) -> None:
         os.chdir(self.dockerfiles_host_dir())
         self.docker.compose.up(detach=True)
@@ -542,15 +719,23 @@ class Installer:
         )
 
     def run_crate_command_and_output_to_file(
-        self, crate_command: str, filename: str
+        self, crate_command: List[str], filename: str
     ) -> None:
-        stdout = self.run_crate_command(crate_command)
-        with open(filename, "w") as f:
-            f.write(stdout)
+        with open(filename, "wb") as f:
+            output_generator = self.run_crate_command(
+                crate_command, stream=True
+            )
+            for stream_type, stream_content in output_generator:
+                if stream_type == "stdout":
+                    f.write(stream_content)
+                elif stream_type == "stderr":
+                    decoded = stream_content.decode("utf-8")
+                    print(decoded, file=sys.stderr, end="")
 
     def run_crate_command(
         self,
-        crate_command: str,
+        crate_command: List[str],
+        stream: bool = False,
         tty: bool = False,
     ) -> Union[str, Container, Iterable[Tuple[str, bytes]]]:
         # Run a command in a new instance of the crate_workers container.
@@ -558,13 +743,24 @@ class Installer:
         # virtualenv or call /bin/bash.
         # "Run" here means "without a terminal".
         if not crate_command:
-            sys.exit("Error: no command specified")
+            self.error("Error: no command specified")
+            sys.exit(EXIT_USER)
+
+        if tty and stream:
+            # Mirror behaviour of docker.compose.run in Python on Whales
+            raise ValueError(
+                "You can't set tty=True and stream=True at the same"
+                "time. Their purpose are not compatible."
+            )
+
         os.chdir(self.dockerfiles_host_dir())
+
         return self.docker.compose.run(
             DockerComposeServices.CRATE_WORKERS,
             remove=True,
+            stream=stream,
             tty=tty,
-            command=[crate_command],
+            command=crate_command,
         )
 
     def exec_crate_command(
@@ -590,29 +786,91 @@ class Installer:
     # Info messages
     # -------------------------------------------------------------------------
 
-    @staticmethod
-    def report(text: str, style: Style) -> None:
-        lines = text.split("\n")
-        wrapped = "\n".join([textwrap.fill(line, width=80) for line in lines])
-        print_formatted_text(HTML(f"<span>{wrapped}</span>"), style=style)
+    def report(
+        self, text: str, style: Style, split_lines: bool = True
+    ) -> None:
+        span = self.span(text, split_lines=split_lines)
+        print_formatted_text(HTML(span), style=style)
 
     def start_message(self) -> None:
         self.report("CRATE Installer", self.intro_style)
+        if self.light_mode:
+            return self.report(
+                "Running in light mode. "
+                "Remove the --light_mode option to run in dark mode.",
+                self.get_highlight_style(),
+            )
+
+        self.report(
+            "Running in dark mode. "
+            "Use the --light_mode option to run in light mode.",
+            self.get_highlight_style(),
+        )
+
+    def highlight(self, text: str) -> None:
+        self.report(text, self.get_highlight_style())
 
     def info(self, text: str) -> None:
         self.report(text, self.info_style)
+
+    def success(self, text: str) -> None:
+        self.report(text, self.success_style)
 
     def envvar_info(self, text: str) -> None:
         if not self.verbose:
             return
         self.report(text, self.envvar_style)
 
-    def error(self, text: str) -> None:
-        self.report(text, self.error_style)
+    def error(self, text: str, split_lines: bool = True) -> None:
+        self.report(text, self.error_style, split_lines=split_lines)
 
     def fail(self, text: str) -> NoReturn:
         self.error(text)
-        sys.exit(EXIT_FAILURE)
+        sys.exit(EXIT_USER)
+
+    def dump_colours(self) -> None:
+        # Development only
+        colour_attrs = inspect.getmembers(
+            Colours, lambda attr: not inspect.isroutine(attr)
+        )
+
+        for name, value in colour_attrs:
+            if name.startswith("__"):
+                continue
+
+            style = Style.from_dict(
+                {
+                    "span": value,
+                }
+            )
+
+            self.report(name, style)
+
+        self.report(
+            "Dark body",
+            Style.from_dict(
+                {"span": f"bg:{Colours.BASE03} fg:{Colours.BASE0}"}
+            ),
+        )
+        self.report(
+            "Dark highlights",
+            Style.from_dict(
+                {"span": f"bg:{Colours.BASE02} fg:{Colours.BASE1}"}
+            ),
+        )
+
+        self.report(
+            "Light body",
+            Style.from_dict(
+                {"span": f"bg:{Colours.BASE3} fg:{Colours.BASE00}"}
+            ),
+        )
+        self.report(
+            "Light highlights",
+            Style.from_dict(
+                {"span": f"bg:{Colours.BASE2} fg:{Colours.BASE01}"}
+            ),
+        )
 
     # -------------------------------------------------------------------------
     # Installation
@@ -673,7 +931,7 @@ class Installer:
     def abort_installation(self) -> None:
         self.error("Installation aborted")
         self.write_environment_variables()
-        sys.exit(EXIT_FAILURE)
+        sys.exit(EXIT_USER)
 
     def configure_user(self) -> None:
         self.setenv(
@@ -760,62 +1018,19 @@ class Installer:
         )
 
         if self.should_create_crate_db_container():
-            return self.configure_crate_db_container()
+            self.configure_crate_db_container()
 
-        self.configure_external_crate_db()
+        self.configure_external_db(self.databases["crate"])
 
     def configure_crate_db_container(self) -> None:
-        self.setenv(InstallerEnvVar.CRATE_DB_ENGINE, "mysql")
-        self.setenv(InstallerEnvVar.CRATE_DB_SERVER, "crate_db")
-        self.setenv(InstallerEnvVar.CRATE_DB_PORT, Ports.MYSQL)
-        self.setenv(
-            DockerEnvVar.CRATE_DB_ROOT_PASSWORD,
-            self.get_docker_crate_db_root_password,
-            obscure=True,
-        )
-        self.setenv(DockerEnvVar.CRATE_DB_DATABASE_NAME, "crate_web_db")
-        self.setenv(DockerEnvVar.CRATE_DB_USER_NAME, "crate_web_user")
-        self.setenv(
-            DockerEnvVar.CRATE_DB_USER_PASSWORD,
-            self.get_docker_crate_db_user_password,
-            obscure=True,
-        )
-        self.setenv(
-            DockerEnvVar.CRATE_DB_HOST_PORT,
-            self.get_docker_crate_db_host_port,
-        )
-
-    def configure_external_crate_db(self) -> None:
-        self.info(
-            "Enter the details of the external database used for the "
-            "CRATE web application"
-        )
-        self.external_db_instructions()
-
-        self.setenv(
-            InstallerEnvVar.CRATE_DB_ENGINE,
-            self.get_external_db_engine,
-        )
-        self.setenv(
-            InstallerEnvVar.CRATE_DB_SERVER,
-            self.get_external_db_server,
-        )
-        self.setenv(
-            InstallerEnvVar.CRATE_DB_PORT,
-            self.get_external_db_port,
-        )
-        self.setenv(
-            DockerEnvVar.CRATE_DB_DATABASE_NAME,
-            self.get_external_db_database_name,
-        )
-        self.setenv(
-            DockerEnvVar.CRATE_DB_USER_NAME,
-            self.get_external_db_user_name,
-        )
-        self.setenv(
-            DockerEnvVar.CRATE_DB_USER_PASSWORD,
-            self.get_external_db_user_password,
-            obscure=True,
+        return self.configure_docker_db(
+            self.databases["crate"],
+            host="crate_db",
+            root_password=self.get_docker_crate_db_root_password,
+            database_name="crate_web_db",
+            user_name="crate_web_user",
+            user_password=self.get_docker_crate_db_user_password,
+            host_port=self.get_docker_crate_db_host_port,
         )
 
     def configure_anon_dbs(self) -> None:
@@ -830,175 +1045,103 @@ class Installer:
             self.configure_demo_secret_db()
             return
 
-        self.configure_external_source_db()
-        self.configure_external_research_db()
-        self.configure_external_secret_db()
+        self.configure_external_db(self.databases["source"])
+        self.configure_external_db(self.databases["research"])
+        self.configure_external_db(self.databases["secret"])
 
     def configure_demo_research_db(self) -> None:
-        self.setenv(InstallerEnvVar.RESEARCH_DATABASE_ENGINE, "mysql")
-        self.setenv(InstallerEnvVar.RESEARCH_DATABASE_HOST, "research_db")
-        self.setenv(InstallerEnvVar.RESEARCH_DATABASE_PORT, Ports.MYSQL)
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_ROOT_PASSWORD,
-            "research",
-            obscure=True,
-        )
-        self.setenv(DockerEnvVar.RESEARCH_DATABASE_NAME, "research")
-        self.setenv(DockerEnvVar.RESEARCH_DATABASE_USER_NAME, "research")
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_USER_PASSWORD,
-            "research",
-            obscure=True,
-        )
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_HOST_PORT, Ports.RESEARCH_DB_HOST
+        self.configure_docker_db(
+            self.databases["research"],
+            host="research_db",
+            root_password="research",
+            database_name="research",
+            user_name="research",
+            user_password="research",
+            host_port=Ports.DEMO_RESEARCH_DB_HOST,
         )
 
     def configure_demo_secret_db(self) -> None:
-        self.setenv(InstallerEnvVar.SECRET_DATABASE_ENGINE, "mysql")
-        self.setenv(InstallerEnvVar.SECRET_DATABASE_HOST, "secret_db")
-        self.setenv(InstallerEnvVar.SECRET_DATABASE_PORT, Ports.MYSQL)
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_ROOT_PASSWORD, "secret", obscure=True
-        )
-        self.setenv(DockerEnvVar.SECRET_DATABASE_NAME, "secret")
-        self.setenv(DockerEnvVar.SECRET_DATABASE_USER_NAME, "secret")
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_USER_PASSWORD, "secret", obscure=True
-        )
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_HOST_PORT, Ports.SECRET_DB_HOST
+        self.configure_docker_db(
+            self.databases["secret"],
+            host="secret_db",
+            root_password="secret",
+            database_name="secret",
+            user_name="secret",
+            user_password="secret",
+            host_port=Ports.DEMO_SECRET_DB_HOST,
         )
 
     def configure_demo_source_db(self) -> None:
-        self.setenv(InstallerEnvVar.SOURCE_DATABASE_ENGINE, "mysql")
-        self.setenv(InstallerEnvVar.SOURCE_DATABASE_HOST, "source_db")
-        self.setenv(InstallerEnvVar.SOURCE_DATABASE_PORT, Ports.MYSQL)
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_ROOT_PASSWORD, "source", obscure=True
-        )
-        self.setenv(DockerEnvVar.SOURCE_DATABASE_NAME, "source")
-        self.setenv(DockerEnvVar.SOURCE_DATABASE_USER_NAME, "source")
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_USER_PASSWORD, "source", obscure=True
-        )
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_HOST_PORT, Ports.SOURCE_DB_HOST
+        self.configure_docker_db(
+            self.databases["source"],
+            host="source_db",
+            root_password="source",
+            database_name="source",
+            user_name="source",
+            user_password="source",
+            host_port=Ports.DEMO_SOURCE_DB_HOST,
         )
 
-    def configure_external_research_db(self) -> None:
-        self.info(
-            "Enter the details of an external research database. This is "
-            "where the anonymised data will be stored. More "
-            "research databases can be added by manually editing the "
-            "configuration files generated by CRATE after installation."
+    def configure_docker_db(
+        self,
+        db: Database,
+        host: Union[str, Callable[[], str]],
+        root_password: Union[str, Callable[[], str]],
+        database_name: Union[str, Callable[[], str]],
+        user_name: Union[str, Callable[[], str]],
+        user_password: Union[str, Callable[[], str]],
+        host_port: Union[str, Callable[[], str]],
+    ):
+        self.setenv(db.engine_var, "mysql")
+        self.setenv(db.host_var, host)
+        self.setenv(db.port_var, Ports.MYSQL)
+        self.setenv(db.root_password_var, root_password, obscure=True)
+        self.setenv(db.name_var, database_name)
+        self.setenv(db.user_name_var, user_name)
+        self.setenv(db.password_var, user_password, obscure=True)
+        self.setenv(db.host_port_var, host_port)
+
+    def configure_external_db(self, db: Database) -> None:
+        if self.external_db_configured(db):
+            return
+
+        self.report(
+            f"\nEnter the details of the {db.short_desc}. {db.long_desc}",
+            self.prompt_style,
         )
         self.external_db_instructions()
 
+        self.setenv(db.engine_var, self.get_external_db_engine)
+        self.setenv(db.host_var, self.get_external_db_server)
+        self.setenv(db.port_var, self.get_external_db_port)
+        self.setenv(db.name_var, self.get_external_db_database_name)
+        self.setenv(db.user_name_var, self.get_external_db_user_name)
         self.setenv(
-            InstallerEnvVar.RESEARCH_DATABASE_ENGINE,
-            self.get_external_db_engine,
-        )
-        self.setenv(
-            InstallerEnvVar.RESEARCH_DATABASE_HOST,
-            self.get_external_db_server,
-        )
-        self.setenv(
-            InstallerEnvVar.RESEARCH_DATABASE_PORT,
-            self.get_external_db_port,
-        )
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_NAME,
-            self.get_external_db_database_name,
-        )
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_USER_NAME,
-            self.get_external_db_user_name,
-        )
-        self.setenv(
-            DockerEnvVar.RESEARCH_DATABASE_USER_PASSWORD,
+            db.password_var,
             self.get_external_db_user_password,
             obscure=True,
         )
 
-    def configure_external_secret_db(self) -> None:
-        self.info(
-            "Enter the details of an external secret administrative database. "
-            "This stores information like Patient ID to Research ID mapping. "
-            "More secret databases can be added by manually editing the "
-            "configuration files generated by CRATE after installation. There "
-            "should be one secret administrative database for every research "
-            "database."
-        )
-        self.external_db_instructions()
+    def external_db_configured(self, db: Database) -> bool:
+        for var in [
+            db.engine_var,
+            db.host_var,
+            db.port_var,
+            db.name_var,
+            db.user_name_var,
+            db.password_var,
+        ]:
+            if os.getenv(var) is None:
+                return False
 
-        self.setenv(
-            InstallerEnvVar.SECRET_DATABASE_ENGINE,
-            self.get_external_db_engine,
-        )
-        self.setenv(
-            InstallerEnvVar.SECRET_DATABASE_HOST,
-            self.get_external_db_server,
-        )
-        self.setenv(
-            InstallerEnvVar.SECRET_DATABASE_PORT,
-            self.get_external_db_port,
-        )
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_NAME,
-            self.get_external_db_database_name,
-        )
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_USER_NAME,
-            self.get_external_db_user_name,
-        )
-        self.setenv(
-            DockerEnvVar.SECRET_DATABASE_USER_PASSWORD,
-            self.get_external_db_user_password,
-            obscure=True,
-        )
-
-    def configure_external_source_db(self) -> None:
-        self.info(
-            "Enter the details of an external source database. This "
-            "contains the data to be anonymised. More "
-            "source databases can be added by manually editing the "
-            "configuration files generated by CRATE after installation."
-        )
-        self.external_db_instructions()
-
-        self.setenv(
-            InstallerEnvVar.SOURCE_DATABASE_ENGINE,
-            self.get_external_db_engine,
-        )
-        self.setenv(
-            InstallerEnvVar.SOURCE_DATABASE_HOST,
-            self.get_external_db_server,
-        )
-        self.setenv(
-            InstallerEnvVar.SOURCE_DATABASE_PORT,
-            self.get_external_db_port,
-        )
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_NAME,
-            self.get_external_db_database_name,
-        )
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_USER_NAME,
-            self.get_external_db_user_name,
-        )
-        self.setenv(
-            DockerEnvVar.SOURCE_DATABASE_USER_PASSWORD,
-            self.get_external_db_user_password,
-            obscure=True,
-        )
+        return True
 
     def external_db_instructions(self) -> None:
         self.info(
             "CRATE will attempt to connect to the external database during "
             "installation."
         )
-        self.info("Before continuing:")
+        self.highlight("Before continuing:")
         self.info(
             "1. The database server must allow remote connections "
             "(e.g. for MySQL bind-address = 0.0.0.0 in mysqld.cnf)."
@@ -1105,7 +1248,7 @@ class Installer:
             self.info(f"Creating {settings}")
             Path(settings).touch()
             self.run_crate_command_and_output_to_file(
-                "crate_print_demo_crateweb_config --leave_placeholders",
+                ["crate_print_demo_crateweb_config", "--leave_placeholders"],
                 settings,
             )
         self.configure_local_settings()
@@ -1240,7 +1383,7 @@ class Installer:
             self.info(f"Creating {config}")
             Path(config).touch()
             self.run_crate_command_and_output_to_file(
-                "crate_anon_demo_config --leave_placeholders", config
+                ["crate_anon_demo_config", "--leave_placeholders"], config
             )
         self.configure_anon_config()
 
@@ -1250,30 +1393,9 @@ class Installer:
             "per_table_patient_id_encryption_phrase": self.get_hmac_md5_key(),
             "master_patient_id_encryption_phrase": self.get_hmac_md5_key(),
             "change_detection_encryption_phrase": self.get_hmac_md5_key(),
-            "dest_db_url": self.get_sqlalchemy_url(
-                self.getenv(InstallerEnvVar.RESEARCH_DATABASE_ENGINE),
-                self.getenv(DockerEnvVar.RESEARCH_DATABASE_USER_NAME),
-                self.getenv(DockerEnvVar.RESEARCH_DATABASE_USER_PASSWORD),
-                self.getenv(InstallerEnvVar.RESEARCH_DATABASE_HOST),
-                self.getenv(InstallerEnvVar.RESEARCH_DATABASE_PORT),
-                self.getenv(DockerEnvVar.RESEARCH_DATABASE_NAME),
-            ),
-            "admin_db_url": self.get_sqlalchemy_url(
-                self.getenv(InstallerEnvVar.SECRET_DATABASE_ENGINE),
-                self.getenv(DockerEnvVar.SECRET_DATABASE_USER_NAME),
-                self.getenv(DockerEnvVar.SECRET_DATABASE_USER_PASSWORD),
-                self.getenv(InstallerEnvVar.SECRET_DATABASE_HOST),
-                self.getenv(InstallerEnvVar.SECRET_DATABASE_PORT),
-                self.getenv(DockerEnvVar.SECRET_DATABASE_NAME),
-            ),
-            "source_db1_url": self.get_sqlalchemy_url(
-                self.getenv(InstallerEnvVar.SOURCE_DATABASE_ENGINE),
-                self.getenv(DockerEnvVar.SOURCE_DATABASE_USER_NAME),
-                self.getenv(DockerEnvVar.SOURCE_DATABASE_USER_PASSWORD),
-                self.getenv(InstallerEnvVar.SOURCE_DATABASE_HOST),
-                self.getenv(InstallerEnvVar.SOURCE_DATABASE_PORT),
-                self.getenv(DockerEnvVar.SOURCE_DATABASE_NAME),
-            ),
+            "dest_db_url": self.get_db_url(self.databases["research"]),
+            "admin_db_url": self.get_db_url(self.databases["secret"]),
+            "source_db1_url": self.get_db_url(self.databases["source"]),
             "source_db1_ddgen_include_fields": "Note.note",
             "source_db1_ddgen_scrubsrc_patient_fields": self.format_multiline(
                 ("forename", "surname")
@@ -1296,43 +1418,41 @@ class Installer:
         )
 
     def create_or_update_crate_database(self) -> None:
-        self.run_crate_command("crate_django_manage migrate")
+        self.run_crate_command(["crate_django_manage", "migrate"], tty=True)
 
     def collect_static(self) -> None:
-        self.run_crate_command("crate_django_manage collectstatic --no-input")
+        self.run_crate_command(
+            ["crate_django_manage", "collectstatic", "--no-input"], tty=True
+        )
 
     def populate(self) -> None:
-        self.run_crate_command("crate_django_manage populate")
+        self.run_crate_command(["crate_django_manage", "populate"], tty=True)
 
     def create_superuser(self) -> None:
         # Will either create a superuser or update an existing one
         # with the given username
-        self.run_crate_command("crate_django_manage ensuresuperuser")
+        self.run_crate_command(
+            ["crate_django_manage", "ensuresuperuser"], tty=True
+        )
 
     def create_demo_data(self) -> None:
-        engine = self.getenv(InstallerEnvVar.SOURCE_DATABASE_ENGINE)
-        user = self.getenv(DockerEnvVar.SOURCE_DATABASE_USER_NAME)
-        password = self.getenv(DockerEnvVar.SOURCE_DATABASE_USER_PASSWORD)
-        host = self.getenv(InstallerEnvVar.SOURCE_DATABASE_HOST)
-        port = self.getenv(InstallerEnvVar.SOURCE_DATABASE_PORT)
-        name = self.getenv(DockerEnvVar.SOURCE_DATABASE_NAME)
-        url = self.get_sqlalchemy_url(engine, user, password, host, port, name)
-        self.run_crate_command(f"crate_make_demo_database {url}")
+        url = self.get_db_url(self.databases["source"])
+        self.run_crate_command(["crate_make_demo_database", url], tty=True)
 
     def create_data_dictionary(self) -> None:
         self.info("Creating data dictionary...")
         data_dictionary = self.get_data_dictionary_host_filename()
         self.run_crate_command_and_output_to_file(
-            "crate_anon_draft_dd", data_dictionary
+            ["crate_anon_draft_dd"], data_dictionary
         )
 
     def anonymise_demo_data(self) -> None:
         self.info("Anonymising demo data...")
-        self.run_crate_command("crate_anonymise --full")
+        self.run_crate_command(["crate_anonymise", "--full"], tty=True)
 
     def report_status(self) -> None:
         localhost_url = self.get_crate_server_localhost_url()
-        self.info(f"The CRATE application is running at {localhost_url}")
+        self.success(f"The CRATE application is running at {localhost_url}")
 
     # -------------------------------------------------------------------------
     # Fetching information from environment variables or statically
@@ -1343,8 +1463,8 @@ class Installer:
 
     def get_docker_install_group_id(self) -> str:
         self.info(
-            "The CRATE Docker image will be created with your user ID and one "
-            "of your user's group IDs so that file permissions will be "
+            "\nThe CRATE Docker image will be created with your user ID and "
+            "one of your user's group IDs so that file permissions will be "
             "correct for any file systems shared between the host and "
             "container."
         )
@@ -1373,10 +1493,12 @@ class Installer:
             return group_id
 
         return self.get_user_choice(
-            "Select the group to use. If a mounted file system needs to "
-            "be shared between multiple users, choose a group that includes "
-            "all of those users.",
             choice_dict,
+            intro_text=(
+                "Select the group to use. If a mounted file system needs to "
+                "be shared between multiple users, choose a group that "
+                "includes all of those users."
+            ),
         )
 
     @staticmethod
@@ -1414,6 +1536,16 @@ class Installer:
     @staticmethod
     def get_data_dictionary_docker_filename() -> str:
         return os.path.join(DockerPath.CONFIG_DIR, "data_dictionary.tsv")
+
+    def get_db_url(self, db: Database) -> str:
+        return self.get_sqlalchemy_url(
+            self.getenv(db.engine_var),
+            self.getenv(db.user_name_var),
+            self.getenv(db.password_var),
+            self.getenv(db.host_var),
+            self.getenv(db.port_var),
+            self.getenv(db.name_var),
+        )
 
     def get_sqlalchemy_url(
         self,
@@ -1521,9 +1653,13 @@ class Installer:
         return os.path.join(self.docker_host_dir(), "dockerfiles")
 
     def docker_host_dir(self) -> str:
-        installer_dir = os.path.dirname(os.path.realpath(__file__))
+        return os.path.join(self.src_host_dir(), "docker")
 
-        return os.path.join(installer_dir, os.pardir, "docker")
+    def src_host_dir(self) -> str:
+        return os.path.join(self.installer_host_dir(), os.pardir)
+
+    def installer_host_dir(self) -> str:
+        return os.path.dirname(os.path.realpath(__file__))
 
     def default_config_host_dir(self) -> str:
         return os.path.join(self.crate_root_host_dir(), "config")
@@ -1620,7 +1756,7 @@ class Installer:
             name: e.description for (name, e) in self.engines.items()
         }
 
-        return self.get_user_choice("Engine:", choice_dict)
+        return self.get_user_choice(choice_dict, prompt_text="Engine:")
 
     def get_external_db_server(self) -> str:
         return self.get_user_optional_input(
@@ -1698,7 +1834,12 @@ class Installer:
     def get_user_optional_input(self, text: str, default: str = "") -> str:
         return self.prompt(text, default=default)
 
-    def get_user_choice(self, text: str, choice_dict: Dict) -> str:
+    def get_user_choice(
+        self,
+        choice_dict: Dict,
+        intro_text: str = None,
+        prompt_text: str = "Enter:",
+    ) -> str:
         definitions_html = "".join(
             [
                 f"<name>{name}</name> for <desc>{description}</desc>\n"
@@ -1709,8 +1850,15 @@ class Installer:
         # noinspection PyTypeChecker
         completer = WordCompleter(choice_dict.keys())
 
+        lines = []
+        if intro_text is not None:
+            lines.append(f"{intro_text}")
+
+        lines.append(prompt_text)
+        span = self.span("\n\n".join(lines))
+
         return self.prompt_html(
-            HTML(f"<span>{text}</span>\nEnter:\n{definitions_html}\n"),
+            HTML(f"\n{span}\n{definitions_html}\n"),
             validator=ChoiceValidator(choice_dict.keys()),
             completer=completer,
             style=self.choice_style,
@@ -1720,12 +1868,28 @@ class Installer:
         """
         Shows a prompt and returns user input.
         """
+        span = self.span(f"\n{text}")
         return self.prompt_html(
-            HTML(f"\n<span>{text}</span> "),
+            HTML(f"{span} "),
             *args,
             **kwargs,
             style=self.prompt_style,
         )
+
+    def span(
+        self, text: str, split_lines: bool = True, escape: bool = True
+    ) -> str:
+        """
+        Returns span HTML fragment, with options to escape and split lines.
+        """
+        if split_lines:
+            lines = text.split("\n")
+            text = "\n".join([textwrap.fill(line, width=80) for line in lines])
+
+        if escape:
+            text = html.escape(text)
+
+        return f"<span>{text}</span>"
 
     @staticmethod
     def prompt_html(html: Union[str, HTML], *args, **kwargs) -> str:
@@ -1820,11 +1984,10 @@ class Installer:
             with NamedTemporaryFile(delete=False, mode="w") as f:
                 envvar_unset_file = f.name
                 self._write_envvar_unsets_to_file(f)
-        self.info(
-            "Settings have been saved and can be loaded with:\n"
-            f"source {envvar_save_file}"
-        )
-        self.info(f"To unset all settings:\nsource {envvar_unset_file}")
+        self.highlight("Settings have been saved and can be loaded with:")
+        self.info(f"source {envvar_save_file}")
+        self.highlight("To unset all settings:")
+        self.info(f"source {envvar_unset_file}")
 
     # -------------------------------------------------------------------------
     # Shell handling
@@ -1886,7 +2049,7 @@ class NativeLinuxInstaller(Installer):
     def report_status(self) -> None:
         server_url = self.get_crate_server_url()
         localhost_url = self.get_crate_server_localhost_url()
-        self.info(
+        self.success(
             f"The CRATE application is running at {server_url} "
             f"or {localhost_url}"
         )
@@ -1929,14 +2092,21 @@ def get_installer_class() -> Type[Installer]:
         return MacOsInstaller
 
     if sys_info.system == "Windows":
-        sys.exit(
+        print(
             "The installer cannot be run under native Windows. Please "
             "install Windows Subsystem for Linux 2 (WSL2) and run the "
             "installer from there. Alternatively follow the instructions "
-            "to install CRATE manually."
+            "to install CRATE manually.",
+            file=sys.stderr,
         )
+        sys.exit(EXIT_USER)
 
-    sys.exit(f"Sorry, the installer can't be run under {sys_info.system}.")
+    print(
+        f"Sorry, the installer can't be run under {sys_info.system}.",
+        file=sys.stderr,
+    )
+
+    sys.exit(EXIT_USER)
 
 
 # =============================================================================
@@ -1962,6 +2132,12 @@ def main() -> None:
             "Top level CRATE directory containing config files and source "
             "code (if not running the installer locally)"
         ),
+    )
+    parser.add_argument(
+        "--light_mode",
+        action="store_true",
+        default=False,
+        help="Use this if your terminal has a light background",
     )
     parser.add_argument(
         "--update",
@@ -2027,6 +2203,7 @@ def main() -> None:
 
     installer = get_installer_class()(
         crate_root_dir=args.crate_root_dir,
+        light_mode=args.light_mode,
         update=args.update,
         verbose=args.verbose,
     )